refactor: move install-ses-debug to internal
